refactor: presentation 모듈 패키지 구조 변경
- common 패키지 추가하여 BaseActivity, BaseFragment 이동
- extension 패키지 common 패키지 하위로 이동
- util 패키지 하위의 permission 패키지 util로 통합
- Event.kt 삭제
