+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/common/extension/DateTimeExtension.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.extension
+package com.bottari.presentation.common.extension
 
 import java.time.LocalDate
 import java.time.LocalTime
