+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/profile/ProfileViewModel.kt
@@ -13,7 +13,7 @@ import com.bottari.di.UseCaseProvider
 import com.bottari.domain.usecase.member.CheckRegisteredMemberUseCase
 import com.bottari.domain.usecase.member.SaveMemberNicknameUseCase
 import com.bottari.presentation.common.event.SingleLiveEvent
-import com.bottari.presentation.extension.update
+import com.bottari.presentation.common.extension.update
 import kotlinx.coroutines.launch
 
 class ProfileViewModel(
