+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/market/detail/MarketBottariDetailFragment.kt
@@ -5,9 +5,9 @@ import android.view.View
 import androidx.core.view.isVisible
 import androidx.fragment.app.viewModels
 import com.bottari.presentation.R
-import com.bottari.presentation.base.BaseFragment
+import com.bottari.presentation.common.base.BaseFragment
+import com.bottari.presentation.common.extension.getSSAID
 import com.bottari.presentation.databinding.FragmentMarketBottariDetailBinding
-import com.bottari.presentation.extension.getSSAID
 import com.bottari.presentation.view.edit.personal.PersonalBottariEditActivity
 import com.bottari.presentation.view.market.detail.adapter.MarketBottariDetailAdapter
